styled component refactored
