+++ src/components/Basic/Dashboard.js
@@ -1,6 +1,6 @@
 import React from "react";
 import { Link } from "react-router-dom";
-import { Button } from "../styles";
+import { Button } from "../../styles/componentStyles";
 
 const Dashboard = (props) => (
   <div>
