+++ src/components/Basic/PageHeader.js
@@ -1,5 +1,5 @@
 import React from "react";
-import { H1 } from "../styles";
+import { H1 } from "../../styles/componentStyles";
 
 const PageHeader = () => (
   <div>
