+++ src/components/Contacts/AddForm.js
@@ -1,7 +1,7 @@
 import React, { useState } from "react";
 import { addContact } from "../../redux/store";
 import { useDispatch } from "react-redux";
-import { Button, Input, Form } from "../styles";
+import { Button, Input, Form } from "../../styles/componentStyles";
 
 export default (AddForm) => {
   let [name, setName] = useState();
