+++ src/index.js
@@ -3,13 +3,16 @@ import ReactDOM from "react-dom";
 import App from "./App";
 import "./index.css";
 import { Provider } from "react-redux";
+import { GlobalStyle } from "./styles/globalStyles";
 import { store } from "./redux/store";
 
 ReactDOM.render(
   <React.StrictMode>
     <Provider store={store}>
+      <GlobalStyle />
       <App />
     </Provider>
   </React.StrictMode>,
   document.getElementById("root")
+  
 );
